Refactor OpenDRIVE speed/tag parsing and split Network-level tests
